% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignmentMatrix.R
\name{alignmentMatrix}
\alias{alignmentMatrix}
\title{aligned topics across chains}
\usage{
alignmentMatrix(
  theta,
  ps,
  K,
  iter = 2000,
  chain = 4,
  SampleID_name = "SampleID"
)
}
\arguments{
\item{theta}{array. three dimensional array (iterations * samples * topic).}

\item{ps}{phyloseq object.}

\item{K}{integer. number of topics.}

\item{iter}{integer. number of iteration used in each chain.}

\item{chain}{integer. number of chains used in MCMC.}

\item{SampleID_name}{character. variable name that contains sample IDs.}
}
\value{
matrix. A matrix of topic assignment for each chain.
}
\description{
aligned topics across chains
}
