% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaAligned.R
\name{betaAligned}
\alias{betaAligned}
\title{Align ASVs proportions in topics across chain.}
\usage{
betaAligned(beta, K, aligned, iter = 2000, chain = 4)
}
\arguments{
\item{beta}{array. three dimensional array (iterations * topic * ASVs).}

\item{K}{integer. number of topics.}

\item{aligned}{matrix. A matrix of topic assignment for each chain. output of alignmentMatrix().}

\item{iter}{integer. number of iteration used in each chain.}

\item{chain}{integer. number of chains used in MCMC.}
}
\value{
array. three dimensional array.
}
\description{
Align ASVs proportions in topics across chain.
}
