% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dds_all_paired_K_11.R
\docType{data}
\name{dds_all_paired_K_11}
\alias{dds_all_paired_K_11}
\title{\code{'DESeqDataSet'} object for the differential topic analysis on all paired-samples in psE_BARBI.}
\format{
An object of class \code{'DESeqDataSet'}.
\describe{
\item{baseMean}{mean of the topic abundance}
\item{log2FoldChange}{log fold change of topic abundance in experimental condions tested.}
\item{lfcSE}{Standard error of log fold change.}
\item{stat}{Test statistic value under the null hypothesis.}
\item{pvalue}{P-value for the test.}
\item{padj}{Adjusted P-value.}
}
}
\source{
Data is from Fitzpatrick et al. (2018)
}
\usage{
dds_all_paired_K_11
}
\description{
A By applying DEseq2::results() on \code{'DESeqDataSet'}, we get a data frame with topics on rows and test results on the columns (baseMean, log2FoldChange, lfcSE, stat, pvalue, padj)
}
\examples{
\dontrun{
results(dds_all_paired_K_11)
}
}
\keyword{datasets}
