% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosticsPlot.R
\name{diagnosticsPlot}
\alias{diagnosticsPlot}
\title{Diagnostics plots - ESS and Rhat.}
\usage{
diagnosticsPlot(theta_aligned, beta_aligned)
}
\arguments{
\item{theta_aligned}{array. Three dimensional array returned by thetaAligned().}

\item{beta_aligned}{array. Three dimensional array returned by betaAligned().}
}
\value{
list. ggplots of effective sample size and rhat plots.
}
\description{
Diagnostics plots - ESS and Rhat.
}
