% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffTop.R
\name{diffTop}
\alias{diffTop}
\title{Differential topic analysis}
\usage{
diffTop(
  design = ~pna,
  ps,
  theta_aligned,
  subsetSample = sample_names(psE_BARBI)
)
}
\arguments{
\item{design}{Factors of interests.}

\item{ps}{phyloseq object.}

\item{theta_aligned}{List. returned object of thetaAligned().}

\item{subsetSample}{Character vector. Sample names to be considered.}
}
\value{
DESeq object.
}
\description{
Differential topic analysis
}
