% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffTopAnalysis.R
\name{diffTopAnalysis}
\alias{diffTopAnalysis}
\title{Differential topic analysis}
\usage{
diffTopAnalysis(design = ~pna, ps, theta_aligned, subsetSample = NULL, ...)
}
\arguments{
\item{design}{Factors of interests.}

\item{ps}{phyloseq object.}

\item{theta_aligned}{List. returned object of thetaAligned().}

\item{subsetSample}{Character vector. Sample names to be considered.}

\item{...}{Arguments to DESeq2::DeSeq.}
}
\value{
DESeq object.
}
\description{
Differential topic analysis
}
