% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDAtopicmodel.R
\name{LDAtopicmodel}
\alias{LDAtopicmodel}
\title{Bayesian latent Dirichlet allocation for 16S and metagenomics data.}
\usage{
LDAtopicmodel(stan_data, ...)
}
\arguments{
\item{stan_data}{A list of stan data.}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}.
}
\description{
Bayesian latent Dirichlet allocation for 16S and metagenomics data.
}
