% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelAssessment.R
\name{modelAssessment}
\alias{modelAssessment}
\title{Model assessement for LDA}
\usage{
modelAssessment(
  ps,
  stan.fit,
  iter,
  statistic = "max",
  ASVsIndexToPlot = c(1, 3, 10:14, 19:26, 36, 51:53, 148)
)
}
\arguments{
\item{ps}{phyloseq object.}

\item{stan.fit}{An instance of stanfit.}

\item{iter}{integer. number of iteration used in each chain.}

\item{statistic}{Function name. Use this statistic to assess the goodness of fit.}

\item{ASVsIndexToPlot}{An integer vector. Use to select ASVs to show the goodness of fit in histograms.}
}
\value{

}
\description{
Model assessement for LDA
}
