% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseqTransformation.R
\name{phyloseqTransformation}
\alias{phyloseqTransformation}
\title{Anscombe (1948) transformation}
\usage{
phyloseqTransformation(ps, cc = 0.4)
}
\arguments{
\item{ps}{phyloseq object.}

\item{cc}{Double. a constant used in the Anscombe (1948) transformation if dispersion is <= 2}
}
\value{
phyloseq object. otu table has transformed values of counts.
}
\description{
Anscombe (1948) transformation
}
