% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotASVCircular.R
\name{plotASVCircular}
\alias{plotASVCircular}
\title{Plot the ASVs distribution in each topic.}
\usage{
plotASVCircular(
  ps,
  K,
  iter = 2000,
  chain = 4,
  beta_aligned,
  varnames = c("iterations", "topic", "rsv_ix"),
  value_name = "beta_h",
  taxonomylevel = "Class",
  thresholdASVprop = 0.008
)
}
\arguments{
\item{ps}{phyloseq object.}

\item{K}{integer. number of topics.}

\item{iter}{integer. number of iteration used in each chain.}

\item{chain}{integer. number of chains used in MCMC.}

\item{beta_aligned}{List. returned object of betaAligned().}

\item{varnames}{Character vector. Use to name column of long formatted data frame of ASVs distribution. An example is c("iterations", "topic", "rsv_ix")}

\item{value_name}{Character. Use to name the summary of posterior samples of ASVs proportions.}

\item{taxonomylevel}{Character. Use to label ASVs.}

\item{thresholdASVprop}{Numberic. Use to select ASVs for circular plot.}
}
\value{

}
\description{
Plot the ASVs distribution in each topic.
}
