% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTopicProportion.R
\name{plotTopicProportion}
\alias{plotTopicProportion}
\title{Plot the topic distribution in all specimens.}
\usage{
plotTopicProportion(
  ps,
  theta_aligned,
  K,
  col_names_theta_all = c("iteration", "Sample", "Topic", "topic.dis"),
  chain = 4,
  warmup = iter/2,
  SampleIDVarInphyloseq = "unique_names",
  design = ~pna
)
}
\arguments{
\item{ps}{phyloseq object.}

\item{theta_aligned}{List. returned object of thetaAligned().}

\item{K}{integer. number of topics.}

\item{col_names_theta_all}{Character vector. Column names for long formatted theta. An example is c("iteration", "Sample", "Topic", "topic.dis")}

\item{chain}{integer. number of chains used in MCMC.}

\item{warmup}{Integer. Number of iterations in each chain used for warmup samples.}

\item{SampleIDVarInphyloseq}{Character. SampleID variable name in phyloseq object.}

\item{design}{Factors of interests.}
}
\value{

}
\description{
Plot the topic distribution in all specimens.
}
