% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerswitch.R
\docType{data}
\name{powerswitch}
\alias{powerswitch}
\title{A vector}
\format{
A vector of integers.
\describe{
\item{first element}{number of pvalues (out of 1000) less than 0.05.}
\item{Second element}{number of pvalues (out of 1000) less than 0.01.}
}
}
\source{
Simulated data based on the data from Fitzpatrick et al. (2018)
}
\usage{
powerswitch
}
\description{
Contais the number of pvalues (out of 1000) less than 0.05 and 0.01 for the PERMANOVA in the simulated data with strain switching.
}
\examples{
\dontrun{
print(round(powerswitch[1]/1000,2))
}
}
\keyword{datasets}
