% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_ans.R
\docType{data}
\name{ps_ans}
\alias{ps_ans}
\title{\code{'phyloseq'} object for the Anscombe transformed data published in Fitzpatrick et al. (2018) after removing DNA contamination.}
\format{
An object of class \code{'phyloseq'}.
\describe{
\item{SampleType}{Whether Asteraceae or non-Asteraceae}
\item{pna}{Type of pPNA used universal or Asteraceae-modified}
\item{species_names}{Sample ID}
}
}
\source{
Fitzpatrick et al. (2018)
}
\usage{
ps_ans
}
\description{
A \code{'phyloseq'} class object with 1418 taxa in 86 samples .
}
\examples{
\dontrun{
sample_data(ps_ans)
otu_table(ps_ans)
}
}
\keyword{datasets}
