% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psE_BARBI.R
\docType{data}
\name{psE_BARBI}
\alias{psE_BARBI}
\title{\code{'phyloseq'} object for the data published in Fitzpatrick et al. (2018) after removing DNA contamination.}
\format{
An object of class \code{'phyloseq'}.
\describe{
\item{SampleType}{Whether Asteraceae or non-Asteraceae}
\item{pna}{Type of pPNA used universal or Asteraceae-modified}
\item{species_names}{Sample ID}
}
}
\source{
Fitzpatrick et al. (2018)
}
\usage{
psE_BARBI
}
\description{
A \code{'phyloseq'} class object with 5808 taxa in 86 samples that are from seven rounds of six-fold dilutions (1:1 up to 1:279,936) from standard ZymoBIOMICS microbial community and 10 negative extraction controls.
}
\examples{
\dontrun{
sample_data(psE_BARBI)
otu_table(psE_BARBI)
}
}
\keyword{datasets}
