% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setUpData.R
\name{setUpData}
\alias{setUpData}
\title{Set-up data for latent Dirichlet allocation}
\usage{
setUpData(ps, K, alpha = 0.9, gamma = 0.5)
}
\arguments{
\item{ps}{phyloseq class.}

\item{K}{Integer. Number of topics.}

\item{alpha}{Non-negative numeric. Hyper-parameter alpha is less than one to generate sparse mixtures that are different from each other and avoid generating unrealistic topics.}

\item{gamma}{Non-negative numeric. Hyper-parameter gamma is less than one to generate sparse mixtures of ASVs in each topic.}
}
\value{
List. List of object with number of topics, number of ASVs,number of specimens, count table, hyper parameters for topic and ASVs distribution.
}
\description{
Set-up data for latent Dirichlet allocation
}
