% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thetaAligned.R
\name{thetaAligned}
\alias{thetaAligned}
\title{Align topic proportions in specimens across chain.}
\usage{
thetaAligned(theta, K, aligned, iterUse = 1000, chain = 4)
}
\arguments{
\item{theta}{array. three dimensional array (iterations * topic * specimens).}

\item{K}{integer. number of topics.}

\item{aligned}{aligned matrix. A matrix of topic assignment for each chain. output of alignmentMatrix.}

\item{iterUse}{integer. number of iterations used in each chain after subtracting warm-up samples.}

\item{chain}{integer. number of chains used in MCMC.}
}
\value{
array. three dimensional array.
}
\description{
Align topic proportions in specimens across chain.
}
