% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeResTable.R
\name{writeResTable}
\alias{writeResTable}
\title{Write differential topic analysis results to latex}
\usage{
writeResTable(res, fileN = "fileName")
}
\arguments{
\item{res}{A data frame. The results produced by DESeq2.}

\item{fileN}{String. A name for the file.}
}
\value{
A tex output will be saved with the filename.
}
\description{
Write differential topic analysis results to latex
}
