% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTopicProportion.R
\name{plotTopicProportion}
\alias{plotTopicProportion}
\title{Plot the topic distribution in all specimens.}
\usage{
plotTopicProportion(
  ps,
  theta_aligned,
  K,
  col_names_theta_all = c("iteration", "Sample", "Topic", "topic.dis"),
  chain = 4,
  iterUse = 1000,
  SampleIDVarInphyloseq = "unique_names",
  design = ~pna
)
}
\arguments{
\item{ps}{phyloseq object.}

\item{theta_aligned}{List. returned object of thetaAligned().}

\item{K}{integer. number of topics.}

\item{col_names_theta_all}{Character vector. Column names for long formatted theta. An example is c("iteration", "Sample", "Topic", "topic.dis")}

\item{chain}{integer. number of chains used in MCMC.}

\item{iterUse}{integer. number of iterations used in each chain after subtracting warm-up samples.}

\item{SampleIDVarInphyloseq}{Character. SampleID variable name in phyloseq object.}

\item{design}{Factors of interests.}
}
\value{
A ggplot2 object. Histrogram of median of topic proportion in each sample.
}
\description{
Plot the topic distribution in all specimens.
}
